refactor(rust-loc): split smtp_server (StreamType, MailServer) + security (types) <300 LOC
